+++ Sources/Audiograph/ObjcInterface.swift
@@ -202,18 +202,20 @@ import UIKit
         audiograph.smoothing = .custom(smoothing)
     }
     
-    /// Creates an accessibility-action that can be applied to the chart-view. The action is configured with the given `AudiographLocalizations` and triggers the Audiograph when activated.
+    /// Creates an accessibility-action that can be applied to any view. The action is configured with the given `AudiographLocalizations` and triggers the Audiograph when activated.
     ///
-    /// Rather use `Audiograph.createCustomAccessibilityAction(forView:)` when you have direct access to the view.
-    /// - Parameter dataProvider: The object that is able to deliver the chart data to the Audiograph-System.
+    /// Rather use `Audiograph.createCustomAccessibilityAction(for:)` when you have direct access to the chart-view to make it known to Audiograph, too.
+    /// - Parameter dataProvider: The object that is able to deliver chart data to the Audiograph-System.
     /// - Returns: An action that can be used to populate `accessibilityCustomActions` of a view.
-    @objc public func createCustomAccessibilityAction(withProvider dataProvider: AudiographProvider) -> UIAccessibilityCustomAction {
+    @objc public func createCustomAccessibilityAction(withProvider dataProvider: AudiographProvidable) -> UIAccessibilityCustomAction {
         return audiograph.createCustomAccessibilityAction(using: dataProvider)
     }
     
     /// Creates an accessibility-action that can be applied to the view. The action is configured with the given `AudiographLocalizations` and triggers the Audiograph when activated.
-    /// - Parameter chartView: The view that can deliver chart data and will receive the created accessibility action.
-    /// - Returns: An action that can be used to populate `accessibilityCustomActions` of a view.
+    ///
+    /// Prefer this method over `Audiograph.createCustomAccessibilityAction(using:)` to make the chart-view known to Audiograph.
+    /// - Parameter chartView: The view that can deliver chart data. The returned action may be applied to that view.
+    /// - Returns: An action that can be used to populate `accessibilityCustomActions` of the chart-view.
     @objc public func createCustomAccessibilityAction(forView chartView: AudiographPlayingView) -> UIAccessibilityCustomAction {
         audiograph.createCustomAccessibilityAction(for: chartView)
     }
@@ -227,7 +229,7 @@ import UIKit
         audiograph.play(graphContent: graphContent, completion: completion)
     }
     
-    /// Stops the preprocessing or audio playback.
+    /// Stops the preprocessing and audio playback.
     @objc public func stop() {
         audiograph.stop()
     }