+++ CHANGELOG.md
@@ -5,10 +5,9 @@ Audiograph adheres to [Semantic Versioning](http://semver.org/).
 ## Next Release:
 - When chart view loses accessibility focus playing audiograph is stopped.
 - Added wrapper for interoperability with Objective-C.
-- Changed public initializer from `Audiograph(localizations: localizations)` to `Audiograph(localizationProvider: localizations)`
 - Changed `AudiographPlayable` to `AudiographProvidable` for an object to declare that it can provide chart data.
-- Introduced `AudiographPlayingView` as a view that can provide audiograph data.
-- `AudiographLocalizations`  is a class now in order to interoperate with Objective-C.
+- Introduced `AudiographPlayingView` as a view that can provide Audiograph data.
+- `AudiographLocalizations` is a class now in order to interoperate with Objective-C.
 
 ## 0.5.0 (26.04.2020):
 - Covered edge case where the input data does not contain enough distint y-coordinates and a steady line is played.